Implement special optimized connection pool
The way that APNS currently works is that it Apple will drop connections
right away if data is not sent. In the previous code, Lead Zeppelin
would try to connect to all sockets at once, but this doesn't work
because a message is only sent to one. This lead to a lot of dead socket
reconnects on mostly idle apps.

With the new implementation, we create a connection, send a message
through it, and then add it to the prescribed connection pool. After a
message is sent, the socket will remain open for 24 hours, and then
close. It is not known if these sockets forcibly close after 24 hours,
or if they only close after 24 hours of inactivity.

This information was derived indirectly by digging on Q/A sites
like Stack Overflow. My testing seems to confirm the validity of this.